git-svn-id: svn://mmweb.animal.net.cn/nb_icesheet/trunk@52 ba629b7a-d26c-4841-822e-5ac455a76d44

+++ src/Definitions/GeoLocation.cpp
@@ -7,7 +7,7 @@
 
 #include "GeoLocation.h"
 
-GeoLocation::GeoLocation(double p_longitude, double p_latitude) {
+GeoLocation::GeoLocation(float p_longitude, float p_latitude) {
     longitude = p_longitude;
     latitude = p_latitude;
 
@@ -17,9 +17,9 @@ GeoLocation::~GeoLocation() {
 
 }
 
-double GeoLocation::getLongitude(){
+float GeoLocation::getLongitude(){
     return longitude;
 }
-double GeoLocation::getLatitude(){
+float GeoLocation::getLatitude(){
     return latitude;
 }
